# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.CreateOrderInput:
    properties:
      customerId:
        type: integer
      items:
        items:
          $ref: '#/definitions/service.OrderItemInput'
        type: array
    type: object
...
  service.OrderItemInput:
    properties:
      productId:
        type: integer
      quantity:
        type: integer
      unitPrice:
        description: 分
        type: integer
    type: object
...
info:
  contact: {}
  description: DDD 模块化单体项目 API 文档
  title: dd-frame API
  version: "1.0"
paths:
...
  /order:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建订单参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateOrderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建订单
      tags:
      - Order
  /order/{orderNo}/cancel:
    post:
      parameters:
      - description: 订单编号
        in: path
        name: orderNo
        required: true
        type: string
      - description: 取消原因
        in: formData
        name: reason
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 取消订单
      tags:
      - Order
  /order/{orderNo}/submit:
    post:
      parameters:
      - description: 订单编号
        in: path
        name: orderNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 提交订单
      tags:
      - Order
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"